Almost done but still need to finish example in README.md
