+++ src/tests/mod.rs
@@ -1,3 +1,3 @@
 mod testcases;
 mod constraint_test;
-mod constraint_test_auto;
+mod constraint_test_various;
